Generate Jupyter Notebook w/o papermill (#456)
* Generate Jupyter Notebook w/o papermill

* remove papermill dependency

* convert to path to Path (so it works on windows)

Also, set default __main__ to use sample_data

* fix imports, move template to same dir

* add error handling for incorrect array size

* test geometry utilities

* error handling for rotation around x axis

* Revert
"add error handling for incorrect array size"

* Revert "test geometry utilities"

This reverts commit 364928079a0ac2ef308773ba54043bb13293b774.

---------
