+++ freemocap/core_processes/process_motion_capture_videos/process_recording_headless.py
@@ -3,11 +3,9 @@ from pathlib import Path
 from typing import Optional, Union
 
 from freemocap.core_processes.process_motion_capture_videos.process_recording_folder import process_recording_folder
+from freemocap.data_layer.generate_jupyter_notebook.generate_jupyter_notebook import generate_jupyter_notebook
 from freemocap.data_layer.export_data.blender_stuff.export_to_blender import export_to_blender
 from freemocap.data_layer.export_data.blender_stuff.get_best_guess_of_blender_path import get_best_guess_of_blender_path
-from freemocap.data_layer.export_data.generate_jupyter_notebook.generate_jupyter_notebook import (
-    generate_jupyter_notebook,
-)
 from freemocap.data_layer.recording_models.post_processing_parameter_models import PostProcessingParameterModel
 from freemocap.data_layer.recording_models.recording_info_model import RecordingInfoModel
 from freemocap.system.paths_and_filenames.file_and_folder_names import RECORDING_PARAMETERS_JSON_FILE_NAME
