+++ pyproject.toml
@@ -77,7 +77,6 @@ dependencies = [
     "toml==0.10.2",
     "aniposelib==0.4.3",
     "libsass==0.21.0",
-    "papermill==2.4.0",
     "ipykernel==6.23.1",
     "plotly==5.14.1",
     "pydantic==1.*"
